[csc] Handle CatalogSourceConfig updates
- Introduce an in-memory CatalogSourceConfig UID : Spec
This cache is used to detect when a CR needs to be updated. Note that
this cache is a temporary construct which will be removed when we move
to using the Operator Registry as the data store for CatalogSources.

- Introduce a new Update reconciler
This reconciler sets the phase back to "Configuring" allowing the
ConfigMap and CatalogSource to be updated or deleted and recreated
depending on whether the target namespace changed. The "Configuring"
reconciler has been updated to detect if the respective object is
already present and to create or update it depending on that.

- Replace default controller client with a non-default one
Now that we need to query other namespaces for ConfigMaps and
CatalogSources, we no longer can use the default client as it serves
read requests from the cache which contains objects only from the
namespace the operator is watching. So we have to create our own client
and pass it the manager's scheme which has all our registered types.