[opsrc] Introduce a new phase: "Purging"
Introduce a new phase called "Purging" and a phase reconciler
associated with it. In this phase a given OperatorSource
object is purged. All resource(s) created as a result of
reconciliation are removed in this phase.

The following scenario(s) should trigger this phase:
 a. An admin changes the spec of a given OperatorSource object.
    This warrants for a purge and reconciliation to start anew.
 b. When marketplace operator restarts it needs to rebuild
    the cache for all existing OperatorSource object(s).

Make the following changes:
- Add a new phase called "Purging".
- Add a new reconciler associated with "Purging" that purges
  a given OperatorSource object.
- Change existing updated reconciler to kick off reconciliation anew
  from "Purging" phase.

This will pave the way for "reconcile out-of-sync OperatorSource".