Ensure mockgen is the same version as gomock
Right now we 'go get' the latest version of golang/mock to build
mockgen binary. On the other hand, The gomock package used in test
is imported from the vendor folder. These two sources of golang/mock
can easily be out of sync. This is exactly why the unit test job
in CI is failing.

We need to ensure that both mockgen and gomock are tied to the same
release. Build mockgen from the vendored copy of golang/mock. This
ensures that both are referenced from the same location and thus
are always tied to the same version.