[opsrc] Make configuring reconciler idempotent
Make configuring reconciler for OperatorSource object idempotent.
Right now, if the given CatalogSourceConfig object already exists
then the reconciler assumes that it is already properly configured
and moves on. This may not be true, due to unexpected error we might
have stale CatalogSourceConfig object. Making the configuring
reconciler idempotent ensures that reconciliation is successful
when CatalogSourceConfig already exists.

Do the following to make it idempotent:
- Call Create to cretae the CatalogSourceConfig object.
- If we get an AlreadyExists error, then get the latest version
  of the object.
- Set the fields appropriately and then Call Update.