[install] Add support for installation by CVO
- Move to using "openshift-marketplace" namespace instead of
"openshift-operators"
- Add a "manifests" folder consisting of an ordered set of resources
required by the operator. CVO will apply them in the given order. Note
that the "imagePullPolicy" is set to "IfNotPresent", which differs from
the deployment in the CSV. This is because the image will be replaced by
the CVO as per image-references.
- Update the Dockerfiles to add the "manifests" folder and
io.openshift.release.operator label.
- Update the README to reflect that Marketplace will be running by
default in the "openshift-marketplace" namespace