[opsrc] Fix OperatorSource reconciliation error
OperatorSource reconciler fails in 'Configuring' phase if the
following conditions are true:
 - The CatalogSourceConfig object already exists.
 - It is already owned by the OperatorSource object.

The reconciler appends the OperatorSource object to the
OwnerReference list of the CatalogSourceConfig without checking
whether it is already owned by the same OperatorSource. This causes
the Update API to throw an error.

Solution: Remove the given OperatorSource object from the
OwnerReference list to ensure it is not specified more than once

https://bugzilla.redhat.com/show_bug.cgi?id=1672701