[csc] Fix update sync with Quay
When there are update present in the Quay namespace but no new packages
are being added or removed, the updates are not getting pulled down.
This was due to the deployment spec for the registry pod being updated
only when the list of packages changes. This commit fixes that issue.

https://bugzilla.redhat.com/show_bug.cgi?id=1678906