Bug 1685066: Fix OperatorSource sync
Problem:
Prior to f9d37cc029, as part of syncing OperatorSources, the
associated CatalogSourceConfigs were being deleted and recreated. This
always ensured that we got the latest state from Quay. With that change
we moved to updating CatalogSourceConfigs. Now in the scenario where a
Quay namespace has changed without app-registry repositories being added
or removed but with existing repositories being updated, these updates
will not be reflected on the cluster.

Solution:
The right fix for this would be for the CatalogSourceConfig cache to
become app-registry version aware and to detect that updates are
required not just based on spec changes. In the short term an easy fix
is to drop the CatalogSourceConfig status before updating.

This fixes https://bugzilla.redhat.com/show_bug.cgi?id=1685066