Bug 1685647: Remove unnecessary ClusterRoles
After moving to using grpc CatalogSources, the operator is dealing with
more resources like `Services` and `Deployments`. Initially there were
being created in a different namespace that required a `ClusterRole`.
But the final implementation saw the resources being created in the same
namespace. Given that we create a `Role` and `RoleBinding` for those
resources we are dealing with in our namespace. The only resources we
need a ClusterRole for are `CustomResourceDefinitions`,
`ClusterOperator` and `CatalogSources`.

This fixes https://bugzilla.redhat.com/show_bug.cgi?id=1685647