Properly sync catalogsourceconfigs
Currently, the opsrc poller will update the package list and send a
notification to the catalog syncer if there are changes to any of its
packages. This has several limitations however, primarily that if there
is any service hiccup or an opsrc is deleted, the history of upgrade is
lost. To alleviate that issue, the following has been updated:

- Added a PackageMap to the status of catalogsourceconfig. This is a
mapping between PackageName and current version at the time the
catalogsourceconfig was created.
- Modified the catalogsyncer to include a new method,
SyncRefresh() that pushes an empty notification onto the listening
channel and calls the Trigger()
- Modified the trigger method so that, when this empty notification is
sent, the triggerer will compare all non-datastore catalogsourceconfigs
with the current versions in the datastore and update them if out of
sync.
- We now call the SyncRefresh() method in two cases. First
whenever the operator first comes up to ensure any cache related issues
are resolved. Second, whenever a new opsrc is created.
- Reduced the initial wait from 10 minutes to 1 minute, to account for
the initial resync.
- Made catalogSyncer global so that the SendRefresh()  method can be
called while the opsrc handler handles creation of new operator sources