Bug 1686668: [csc] Ignore invalid values when reading `csc.Spec.Packages`
Problem:
In some cases, when the Openshift UI updates the package list of a csc
after removing a subscription, the update does not properly format the
package list. For example:

`packages: ',strimzi-kafka-operator-test'`

In this example, the package list accidentally includes a leading comma.

Right now, the marketplace doesn't handle poorly formatted package
lists in the csc spec. In general, if a user accidentally leaves a comma
behind the operator should ignore it and proceed as normal with the
list of valid packages.

Solution:
Any time the operator reads the package list, that read happens
from a public method that formats the package list and excludes invalid
packages (specifically empty string).
Added a private method `getValidPackageSliceFromString()` that takes a
package string and removes invalid elements from it.
Updated the two `GetPackages()` methods on the csc and spec types to
call this method before returning.
Added a new method on the csc type called `GetPackages()` that uses the
validation logic to reconstruct the `Spec.Packages` string.
Updated every case where we were referencing the `Spec.Packages` field
directly and replaced it with one of these methods.

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=1686668