Bug 1691241: [csc] Don't ignore cache on restart
Problem:
Rebuilding the `CatalogSourceConfig` state relies on the cache
implementation. Currently, if the spec is updated on a csc it will
not automatically attempt to reconcile that change. Instead, the phase
factory checks if the spec has been updated from the previous update.
To accomplish this, it checks an in-memory cache that compares the
cached spec to the current spec. This is troublesome in the case where
the `marketplace` restarts, because the in memory cache is lost. This
bug is caused when an existing `catalogsourceconfig` gets an update:
the cache is empty for that `csc` but we do not return stale.

Resolution:
Update the `IsEntryStale` method on the `csc` cache. In all cases, if
the `csc` is not found in the cache we should return stale.
Also update `GetPhaseReconciler` to only check the cache is stale when
the phase is not set to `Initial` or `Configuring`, since we are already
in the process of reconciling and updating the cache in that case. This
also prevents infinite loops of constantly going to `Update`.

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=1691241