Bug 1691241 [csc] Prevent getting stuck in current state when cache is invalid for csc
Problem:
When the service restarts, we lose all cached knowledge about the
catalogsourceconfig, including the old targetNamespace. This is an
issue, because if the targetNamespace changed we want to remove the
catalogsource in the old targetNamespace and replace it with a new one.
Now that we are always assuming the cache is invalid if it is empty
(which is a valid assumption), we still don't have a way of cleaning up
old targetNamespace if it changed. Right now though, we will return an
error and not attempt to reconcile the CSC.

Not only that, but this error would also cause the catalogsourceconfig
to get into a stuck state. When the error was returned, we moved back
into the current state to retry. If the state was `success` or `failure`
the `catalogsourceconfig` would simply loop in that phase until the
status was dropped or the syncer set the status to initial.

Solution:
If the cache is invalid we cannot assume that the targetNamespace is
actually invalid. For now, we should just print a debug log and return
and continue to reconcile the csc. Once we have a better method
of retaining references to owned objects, we can handle this edge case
better.

Additionally, update the `deleteObjects` function to no longer return
an error. We do not want to cycle forever attempting to clean up the
`catalogsource` if we are unable to, since that would cause our
`catalogsourceconfig` to be stuck in a bad state looping on the state
until the syncer restarted the reconcilliation loop or through manual
intervention.

This issue was uncovered by the original fix for
https://bugzilla.redhat.com/show_bug.cgi?id=1691241, and this commit
is addition to that fix.