Bug 1693375: [csc] Fix operator-registry command
In the case where a CatalogSourceConfig had packages from two different
app registries, the generated command was getting mangled by the time
it to got to the registry pod. This was fixed creating a list of -r
options that was then passed to the registry pod.

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=1693375