[docs] Reorganize README to make it easier to read
- Remove pre-alpha status
- Changed CRD descriptions from paragraph to bullet form
- Added link to authenticated-private-repo doc in
  Populating AppRegistry section

The README has a lot of information in it, and it's hard for users
to find the information they're looking for.
The Description section has useful info and links, which were
difficult to see or find while it was in paragraph form.
I've added a redundant link to the private AppRegistry repository
doc in the Populating AppRegistry section since this it is related
to that section.