Bug 1700416:[status] Fix Status Reporting
Problem:
When the marketplace-operator starts up, it starts a cluster status monitoring
routine. In some cases, the marketplace-operator will incorrectly mark the
status as failing when a large enough number of operands fail to reconcile
successfully (for example, OperatorSources that are pointing to an invalid
app registry path).

Solution:
Update marketplace-operator to report availability based on whether or not it
is able to reconcile CatalogSourceConfigs and OperatorSources. Report failing
if OperatorSources or CatalogSourceConfigs fall below golden ratio, but keep
available condition set to true.

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=1700416