[status] Rely on stopCh to stop reporting status
Problem:
The routine responsible for ClusterOperator status reporting is stopped
when the monitorStopCh is closed. The monitorStopCh is only closed
when the stopCh is closed and is not used for any other purposes.

Solution:
Rather than creating a channel that is closed when the stopCh channel
is closed, we can simply wait for the stopCh channel to close.