[registry] Create registry, builders, and cs pkgs
Problem:
OperatorSources will directly create datastore CatalogSources. With
this change, the OperatorSource reconciler will need access to the
registry code that is embedded in the catalogsourceconfig package.

Solution:
Move registry code to its own package. Move builders code to its own
package. With the catalogsourceconfigbuilder previously part of the
operatorsource package, some of its content will need to be exported.
Move catalogsource creation to its own package. Change the method
signatures of these registry methods that assume the creation of an
in-memory CatalogSourceConfig (since this will no longer be the case
in the OperatorSource reconciler).