Bug 1717439: [status] Populate RelatedObjects
Problem:
The must-gather tool is not able to gather enough information about the
marketplace operator.

Solution:
The must-gather tool requires a field, RelatedObjects, in the
ClusterOperator CR to be populated with ObjectReferences of the
resources associated with the operator. This PR populates this field
with the operator's namespace and the default
OperatorSource/CatalogSourceConfig/CatalogSource resources.