Bug 1717439: [status] Populate RelatedObjects
Problem:
The must-gather tool is not able to gather enough information about the
marketplace operator.

Solution:
The must-gather tool requires a field, RelatedObjects, in the
ClusterOperator CR to be populated with ObjectReferences of the
resources associated with the operator. This PR populates this field
with the operator's namespace and
OperatorSource/CatalogSourceConfig/CatalogSource resources.

(based on commit 5b7e6fb48a6e8c7d143e39df491590a01d02cd92)