[migration] Update existing Subscriptions and remove stale resources
With #198 introducing a change in a way that a datastore CatalogSource
is directly created by an OperatorSource instead of a CatalogSourceConfig,
this PR performs some upgrade logic to update/cleanup existing/stale
resources in existing clusters. Specifically:
* Subscriptions created by the UI with lables "csc-owner-name"
and "csc-owner-namespace" are updated to reference datastore CatalogSources,
instead of installed CatalogSoures.
* ClusterRole was added to list Subscriptions in all namespaces.
* Stale datastore CatalogSourceConfigs created before #198 are deleted.
* Stale installed CatalogSourceConfigs created during operator installation
in old clusters are deleted.