Bug 1732164:[metrics] Report metrics for default operator-sources
If any of the three deafult operator-sources (redhat, community or
certified) fail due to marketplace not being able to connect to quay.io,
the corresponding metrics consisting of opsrc-name and error
message are reported to prometheus. Further, alerts
are configured for these metrics, so that they can be integrated
with the Telemetry server.