[csc] Ensure cscs not deleted on operator restart
Ensure only CatalogSourceConfigs that were created by an OperatorSource
or by the UI as part of the 4.1 install flow are deleted on operator
restart. CatalogSourceConfigs that do not fall into these two categories
should not be deleted. This fixes the labelSelector used to query such
CatalogSourceConfigs prior to deletion and adds a test case for this
scenario.