Restore default opsrc spec
Problem:
Currently, an issue exists where if a default opsrc is modified and that
modification results in a reconciliation loop that puts the opsrc in a
failed state, it does not recover automatically.

Solution:
Rather than rely on the reconciliation loop to run erroneously on these
opsrces and then returning them to their original state afterwards, add
a filter to prevent updates to the spec that do not match the default.
Do this by adding method to the defaults package `RestoreSpecIfDefault`
that always overwrites the spec of any of the defaults to what is
expected, and call that method at the beginning of the opsrc handler.