Bug 1700100: Stop scaling down registry deployment
Problem: Currently the operator scales down and then scales up the
registry deployment, to cause it to forcefuly relaunch the pod so that
updates from Quay are always picked up even when packages are not added
or removed. This causes issues when we pick up a bad update that results
in a crash looping pod resulting in an invalid registry for the
CatalogSource.

Solution: Rather scaling down and up the deployment, update an
annotation in the pod template spec which will also force the update.

Note: This PR alone won't fix BZ1700100. A fix to the e2e test that
doesn't fail it on crash looping pods in the openshift-marketplace
namespace is required.