[CA] Remove CA support
The marketplace-operator relies on the cluster-network-operator to
inject a ConfigMap with the trusted Certificate Authorities bundle.
The marketplace-operator is upgraded at position 50 while the
cluster-network-operator is upgraded at position 70. Because of the
ordering, the marketplace-operator will not start due to the missing
CA bundle.

Due to the severity of the issue, marketplace is being reverted to the
latest stable version.