Go test timeout fix
e2e tests fail now with the addition of the operator restart test
which checks that default opsrces are recreated (due to the length
of that test). Increasing the default e2e timeout to 20 minutes to
allow the tests to complete.

In the future, these tests should be enhanced. Many of them could
run in parallel with proper consideration.