[status] Wait an hour before reporting degraded
This commit introduces a change that prevents Marketplace from reporting
that is is in a degraded state for an hour if the pod is missing the
certificate authority bundle that should exist on the pod.

This change is introduced due to the fact that the marketplace operator
relies on the cluster-network-operator to inject a ConfigMap with the
certificate authority information for proxy enablement in OpenShift 4.2.
This feature was not available in OpenShift 4.1, and when upgrading from
OpenShift 4.1 to 4.2, the marketplace operator is updated first. As a
result, marketplace will report that it has degraded due to the
combination of a missing CA bundle and failing enabled default
OperatorSources.

If the marketplace operator is still failing to reconcile the enabled
default OperatorSources after an hour, marketplace will report that it
has degraded.