Handle delete events while operator is restarting
Problem:
An edge case bug was added when the default opsrc feature was added.
If a default opsrc was marked for deletion but the finalizer was
not removed by the operator before the operator restarted, then the
`defaults.EnsureAll()` method would attempt to create it. In that case,
the client would return an error and the operator would exit.

Solution:
When EnsureAll is called on startup, also ensure that there are no
finalizers still included on the operator source. If they are still
there, ignore it and allow the normal reconciliation cycle to take care
of deleting and recreating it.

Additionally, once the finalizer is removed at the end of the
reconciliation loop, wait for the opsrc to be cleaned up by the garbage
collector before recreating it.