Race condition in config test
A race condition can occur in the `testDisableNonDefault()` test in
the `OperatorHubTests` test suite e2e tests where the test updates the
config object then immediately does a get without waiting for the config
object's controller to handle the update event. In some cases, this can
cause a test failure because the get returns the old object that is yet
to be updated.

This commit updates the test to poll against the object until the config
object has a chance to be updated before continuing with the test.