[bug] CSC not resolving
Problem:
If a CatalogSourceConfig specifies a source that does not exist the
CatalogSourceConfig will get stuck in the configuring phase. This error
is encountered because the Reconcile method for the CatalogSourceConfig
is not returning an error, causing the handler not to reconcile the
CatalogSourceConfig.

Solution:
Update the CatalogSourceConfig Configuring Reconcile method to return an
error. Also adds e2e tests to prevent regressions in the future.