Assume available status is true without defaults
Problem:
The clusteroperator operator status is never set to Available=True when
there are no default operator sources set. This is because it relies on
a set of reconciliations to return successfully and determines a ratio
based on that value. When no default operator sources are enabled, there
are no reconciliation events taking place and therefore the ratio is
always set to zero.

Solution:
When no default operator sources are enabled and no events are firing,
set the Available status to true.

Additionally, reduce the number of minimum events before reporting
status in the event that only a single default operatorsource is enabled

Note: Integration tests are added for disabling all/some default operators.