Bug 1762006: Assume available status is true without defaults
Problem:
The cluster operator status is never set to Available=True when
there are no default operator sources set. This is because it relies on
a set of reconciliations to return successfully and determines a ratio
based on that value. When no default operator sources are set, there are
no reconciliation events taking place and therefore the ratio is always
set to zero.

Solution:
When no default operator sources are set and no events are firing, set
the Available status to true.

Cherry-picked from master branch

https://bugzilla.redhat.com/show_bug.cgi?id=1762006