Bug 1767547: Don't create invalid deployment
Problem Statement:
Replicasets are being constantly generated when the `targetNamespace`
of a CatalogSourceConfig does not exist on the cluster.
The reconciliation loop relies on the controller's exponential backoff
so that events are not constantly reconciled. However, because the
`EnsureResources()` function recreates the registry service in order to
generate a new ip address before attempting to create the catalogsource,
the previous child resource watch constantly forces a reconciliation.
This causes the failing deployment to update constantly without ever
succeeding, and replicasets are continuously spawned.

Solution:
Ensure that the `targetNamespace` exists before attempting to recreate
the service. If the namespace does not exist, log an error and return.
This will prevent the watch from firing and events will back off. When
the namespace is recreated, the controller will eventually pick up the
event.

https://bugzilla.redhat.com/show_bug.cgi?id=1769841