Use DNS for CatalogSource address
Problem:
Marketplace needs to support ipv6 clusters. Currently when creating
catalogsources, the marketplace points the catsrc to the clusterIP
address of the registry pod that it spins up. This cluster ip will not
work in an ipv6 only cluster.

Solution:
Use the dns address of the service to wire up the catalogsource to the
registry pod.