Add CLI flag to set ClusterOperator name.
If the flag is set to a non-empty string, and config.openshift.io/v1
isn't available, then the operator exits with an error
status. Previously, the absence of configv1 would log a warning and
continue running without ever updating ClusterOperator status. If this
occurred during an OpenShift cluster upgrade, the upgrade would time
out waiting for marketplace to report success via its ClusterOperator.

Outside of an OpenShift cluster, the new flag can be omitted (or set
to the empty string) to disable ClusterOperator status updates
altogether.