Bug 1831818: Remove Obsolete default OperatorSource
In #300, the OperatorHub API was enhanced to accept both OperatorSource
and CatalogSource as defaults. However, when an existing OperatorSource
in a cluster was switched to a CatalogSource, the old OperatorSource
persisted. This PR fixes the issue, and removes the obsolete OperatorSource,
so that there's only one source for a catalog of operators.