Bug 1868770: Allow catsrc with default catsrc name in disconnected env
With the introduction of #336, when a CatalogSource with the same name as that of a
default CatalogSource name was being created, the new CatalogSource was either being
restored to the default Spec, or being removed if the CatalogSource was disabled via
OperatorHub. This made is impossible to create a CatalogSource with the same name as
that of the default CatalogSource in a disconnected cluster enviroment. This PR fixes
the issue by allowing for creation/update/deleteion of a CatalogSource with the same
name as that of a default CatalogSource, if it has been disabled via OperatorHub first.
When the default CatalogSource is renabled via OperatorHub, the default CatalogSource
Spec is restored by overriding the custom created CatalogSource.