scripts: Add the ability to override the operator-sdk binary in e2e tests
Update the e2e-job and e2e-tests bash scripts and fix the issue where
the /tmp/operator-sdk binary isn't in $PATH locally by adding the ability
to override the operator-sdk binary we're testing against.

Update the scripts/e2e-tests.sh script and export GO111MODULE=off now
that we're building and testing with Go 1.16, but we're still using dep
as the package manager.