Bug 1946790: Update clusteroperator status conditions on startup
Previously, the conditions for clusteroperator marketplace was set to
available:false, progressing:true, upgradable:true, degraded:false with the
message "Progressing towards release version..." on startup of the marketplace
operator. This was done because a test in the openshift/origin test suites checked
to make sure operators did not set their condition to available:true immediately
on startup. That test has however been removed.
This PR sets the conditions for clusteroperator marketplace to available:true,
progressing:false, degreaded:false and upgradable:true on startup instead of
waiting for a particular time period.