Bug 1936585: Configure alerts for Non-Ready Default CatalogSources
With the [introduction of the `catalogsource_ready` metric in olm](https://github.com/operator-framework/operator-lifecycle-manager/pull/2152), alerts can be
fired for the default CatalogSources marketplace deploys if they are in a non-ready state.
This PR introduces prometheus alerts for any default CatalogSources that have
been in a Non-Ready state for more than 10 mins.