Start without defaults on ARM
The defaults currently point to 4.8 indexes which do not yet exist on
ARM, which result in unnecessary alerts and the need for a post-install
command to silence them.  Removing the defaultDir argument from the
operator launch command on ARM should avoid this.

This should be reverted once 4.9 indexes are available.