pkg,test: Remove unused packages
Remove the unused pkg/proxy/proxy.go test as it's no longer needed with
the deprecated and removal of the OperatorSource API.

Remove the ProxyTests test handler package. Previously, this test
handler was being registered before the OperatorSource API was removed.

During the removal of the OperatorSource API [1], the removal of this
package was missed.

[1] operator-framework@82b5ef3