Fix operatorhub status
Problem: Within the processCatsrc function a GET call is made to the
API for the catalogSource with a given namespace/name the error is
ignored if a 404 is returned. When defaultCatalogSources are disabled
and a new catalogSource named after one of the default catalogSources
exists it is possible for the previous 404 to be returned when no
error is encountered.

Fix: Do not allow the GET error to be returned at the end of the
processCatsrc function.

Signed-off-by: Alexander Greene <greene.al1991@gmail.com>