Bump of-api and Kubernetes dependencies
This commit updates the marketplace operator to use the latest version
of the operator-framework api, which required a kubernetes dependency
bump. These changes are being introduced so the marketplace-operator
is able to apply some of the latest changes introduced to the
catalogSource api.

This commit also updates the operator marketplace to use the new
coordination.k8s.io/leases api for leader election.