Scope Marketplace Operator RBAC
Problem: The Marketplace Operator historically was responsible for
making content available on cluster and serving it through pods. The
Marketplace Operator has since evolved to simply manage the state of
default catalogSource CRs based on specifications in the OperatorHub
resource.

Solution: Remove any RBAC that is no longer necessary for the
Marketplace Operator to serve its purpose.