Dockerfile: Align OKD and OCP
Alignes the FROM directive in both Dockerfiles with one commonly used elsewhere in OpenShift.
This facilitates builds outside of the Prow CI system.

This change also replaces Dockerfile.rhel7 with a symlink to Dockerfile.
The replacement of Dockerfile.rhel7 is a cleanup that'll allow for removal of Dockerfile.rhel7 eventually.