OCPBUGS-59570, OCPBUGS-59566: Move Network Policy manifests
The manifests are in the wrong place currently. This repository has
static manifests that get copied to the container image during build time,
from a static location (the `/manifests` folder). Moving over the
NetworkPolicy manifests to the right location so that they're actually
applied on cluster.

Also include neccessary annotations that were missing previously, without
which the CVO won't apply the manifests:

```
annotations:
    include.release.openshift.io/self-managed-high-availability: "true"
    include.release.openshift.io/single-node-developer: "true"
    capability.openshift.io/name: "marketplace"
```
Signed-off-by: Anik Bhattacharjee <anbhatta@redhat.com>